geo: Improve distribution of geo_color_for

+++ libs/geo/include/geo_color.h
@@ -48,7 +48,7 @@ ASSERT(alignof(GeoColor) == 16, "GeoColor has to be aligned to 128 bits");
  * Get a (deterministic) color for an index or hash.
  * Useful for generating debug colors.
  */
-GeoColor geo_color_for(u64 idx);
+GeoColor geo_color_for(u32 idx);
 GeoColor geo_color_for_hash(u32 hash);
 
 /**
